+++ R/visualize_pathways_level1.R
@@ -321,7 +321,7 @@ visualize_pathways_level1 <- function(df,
     ))
   # Import all CBD pathways level 1
   pathways_level1_all <-
-    pathways_cbd %>%
+    pathwayscbd %>%
     dplyr::distinct(.data$pathway_level1)
   # Select pathways
   if (!is.null(pathways)) {